Struct icu_pattern::Pattern

pub struct Pattern<'s, P>(_);

Pattern stores the result of parsing operation as a vector of PatternToken elements.

Type parameters

P: The type of the placeholder used as a key for the ReplacementProvider.

Lifetimes

p: The life time of an input string slice to be parsed.

Implementations

impl<'s, P> Pattern<'s, P>

pub fn interpolate<'i, E, R>(
&'i self,
replacements: &'i R
) -> Result<InterpolatedPattern<'i, 's, E>, PatternError<R::Key>> where
R: ReplacementProvider<'i, E, Key = P>,
P: Debug + FromStr + PartialEq + Clone,
<P as FromStr>::Err: Debug + PartialEq,

Interpolates the Pattern with provided replacements and returns a InterpolatedPattern structure.

Example

use icu_pattern::{Pattern, InterpolatedKind};
use std::{
    convert::TryFrom,
    ops::Deref
};

#[derive(Debug, PartialEq)]
struct Element(usize);

let pattern = Pattern::try_from("${0}")
    .expect("Failed to parse a pattern");

let replacements = vec![
    Element(5)
];

let interpolated_pattern = pattern.interpolate(&replacements)
    .expect("Failed to interpolate");

assert_eq!(
    interpolated_pattern.deref(),
    &[
        InterpolatedKind::Literal(&"$".into()),
        InterpolatedKind::Element(&Element(5)),
    ],
);

For allocation-free interpolation, see interpolate_to_string or interpolate_to_write.

For lower level interpolation iterator see Interpolator.

pub fn interpolate_to_string<'i, E, R>(
&'i self,
replacements: &'i R
) -> Result<String, PatternError<R::Key>> where
R: ReplacementProvider<'i, E, Key = P>,
P: Debug + FromStr + PartialEq + Clone,
<P as FromStr>::Err: Debug + PartialEq,
E: 'i + Display,

Interpolates the Pattern with provided replacements and a new String.

Example

use icu_pattern::Pattern;
use std::{
    convert::TryFrom,
    fmt::Display
};

#[derive(Debug, PartialEq)]
struct Element(usize);

impl Display for Element {
    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
        self.0.fmt(f)
    }
}

let pattern = Pattern::try_from("${0}")
    .expect("Failed to parse a pattern");

let replacements = vec![
    Element(5)
];

let interpolated_pattern = pattern.interpolate_to_string(&replacements)
    .expect("Failed to interpolate");

assert_eq!(
    interpolated_pattern,
    "$5",
);

For buffer write interpolation, see interpolate_to_write.

For lower level interpolation iterator see Interpolator.

pub fn interpolate_to_write<'i, E, R, W>(
&'i self,
replacements: &'i R,
sink: &mut W
) -> Result<(), PatternError<R::Key>> where
R: ReplacementProvider<'i, E, Key = P>,
P: Debug + FromStr + PartialEq + Clone,
<P as FromStr>::Err: Debug + PartialEq,
E: 'i + Display,
W: Write,

Interpolates the Pattern writing the result into a buffer.

Example

use icu_pattern::Pattern;
use std::{
    convert::TryFrom,
    fmt::Display
};

#[derive(Debug, PartialEq)]
struct Element(usize);

impl Display for Element {
    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
        self.0.fmt(f)
    }
}

let pattern = Pattern::try_from("${0}")
    .expect("Failed to parse a pattern");

let replacements = vec![
    Element(5)
];

let mut result = String::new();
pattern.interpolate_to_write(&replacements, &mut result)
    .expect("Failed to interpolate");

assert_eq!(
    result,
    "$5",
);

pub fn interpolate_to_writeable<'i, E, R, W>(
&'i self,
replacements: &'i R,
sink: &mut W
) -> Result<(), PatternError<R::Key>> where
R: ReplacementProvider<'i, E, Key = P>,
P: Debug + FromStr + PartialEq + Clone,
<P as FromStr>::Err: Debug + PartialEq,
E: 'i + Writeable,
W: Write,

Interpolates the Pattern writing the result into a Writeable buffer.

Example

use icu_pattern::Pattern;
use writeable::Writeable;
use std::{
    convert::TryFrom,
};

#[derive(Debug, PartialEq)]
struct Element(usize);

impl Writeable for Element {
    fn write_to<W>(&self, sink: &mut W) -> Result<(), std::fmt::Error>
    where
        W: std::fmt::Write + ?Sized,
    {
        sink.write_str(&self.0.to_string())
    }
}

let pattern = Pattern::try_from("${0}")
    .expect("Failed to parse a pattern");

let replacements = vec![
    Element(5)
];

let mut result = String::new();
pattern.interpolate_to_writeable(&replacements, &mut result)
    .expect("Failed to interpolate");

assert_eq!(
    result,
    "$5",
);