icu_pattern

Trait PatternBackend

pub trait PatternBackend:
    Sealed
    + 'static
    + Debug {
    type PlaceholderKey<'a>;
    type PlaceholderKeyCow<'a>;
    type Error<'a>;
}

Types that implement backing data models for Pattern implement this trait.

The trait has no public methods and is not implementable outside of this crate.

๐ซ This trait is sealed; it cannot be implemented by user code. If an API requests an item that implements this trait, please consider using a type from the implementors listed below.

Required Associated Types

type PlaceholderKey<'a>

The type to be used as the placeholder key in code.

type PlaceholderKeyCow<'a>

Cowable version of the type to be used as the placeholder key in code.

type Error<'a>

The type of error that the TryWriteable for this backend can return.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.